requirements.txt
# Core Dependencies
torch>=1.12.0
transformers>=4.21.0
sentence-transformers>=2.2.0
scikit-learn>=1.1.0
pandas>=1.4.0
numpy>=1.21.0

# Clustering and Dimensionality Reduction
umap-learn>=0.5.3
hdbscan>=0.8.28
plotly>=5.10.0

# Data Processing
tqdm>=4.64.0
concurrent-futures>=3.1.1
lmdb>=1.3.0
gzip

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.11
summac>=0.1.0
openai>=0.27.0

# Configuration
pyyaml>=6.0
json5>=0.9.0

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0
pytest-mock>=3.7.0
pytest-benchmark>=4.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.15.0

# GPU Support (optional)
cupy-cuda11x>=10.6.0  # For CUDA 11.x
cupy-cuda12x>=12.0.0  # For CUDA 12.x

# Documentation
mkdocs>=1.4.0
mkdocs-material>=8.5.0
mkdocs-mermaid>=0.6.0
mkdocs-macros>=0.7.0

# Utilities
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
